- XmlData inherited classes are managed through synchronized references counted pointers + pool, to avoid too much heap new/delete operations and to automatically "garbage" no more used instances like XmlEntity (eg. Avatar::XmlEntity attribute is used after Avatar instance deletion to send the lost event to other peers/nodes) - bug fix into peerPhysX plugin on controller manager releasing crash when 2 avatars are connected then the first connected finally disconnect, legacy ::ControllerManager used instead of NxControllerManager
git-svn-id: svn://scm.gforge.inria.fr/svnroot/solipsis/trunk@600 53240bf5-9d7a-4d4b-845a-3b2958f72d47

+++ nodeSimulator/AvatarNode.cpp
@@ -10,14 +10,21 @@ AvatarNode::AvatarNode(const NodeId& nodeId, XmlEntity* xmlEntity) :
     mMutex(PTHREAD_MUTEX_INITIALIZER),
     mAvatar(xmlEntity)
 {
+    LogManager::getSingleton().logMessage("AvatarNode::AvatarNode() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()));
+
 #ifdef PHYSICSPLUGINS
     IPhysicsEngine* engine = PhysicsEngineManager::getSingleton().getSelectedEngine();
+    if (engine == 0)
+        Exception(Exception::ERR_INTERNAL_ERROR,
+        "No physics engine selected !",
+        "AvatarNode::AvatarNode");
+    LogManager::getSingleton().logMessage("AvatarNode::AvatarNode() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " creating physics scene with engine:" + engine->getName());
     // Create the physical scene
     mPhysicsScene = engine->createScene();
     if (!mPhysicsScene->create())
         Exception(Exception::ERR_INTERNAL_ERROR,
         "Unable to create the PhysX scene !",
-        "PhysXScene::PhysXScene");
+        "AvatarNode::AvatarNode");
 #endif
 
     addAwareEntity(&mAvatar);
@@ -28,12 +35,15 @@ AvatarNode::AvatarNode(const NodeId& nodeId, XmlEntity* xmlEntity) :
 //-------------------------------------------------------------------------------------
 AvatarNode::~AvatarNode()
 {
+    removeAwareEntity(&mAvatar);
+
     pthread_mutex_lock(&mMutex);
 
     if (!mFrozen)
         Peer::getSingleton().removeTimeListener(this);
 
 #ifdef PHYSICSPLUGINS
+    LogManager::getSingleton().logMessage("AvatarNode::~AvatarNode() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " destroying physics scene");
     if (mPhysicsScene != 0)
         PhysicsEngineManager::getSingleton().getSelectedEngine()->destroyScene(mPhysicsScene);
     mPhysicsScene = 0;
@@ -65,18 +75,27 @@ IPhysicsScene* AvatarNode::getPhysicsScene()
 //-------------------------------------------------------------------------------------
 bool AvatarNode::addAwareEntity(Entity* entity)
 {
+    LogManager::getSingleton().logMessage("AvatarNode::addAwareEntity() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " adding entity uid:" + StringConverter::toString(entity->getXmlEntity()->getUid()));
+
     pthread_mutex_lock(&mMutex);
 
     mAwareEntities[entity->getXmlEntity()->getUid()] = entity;
     if (entity->getXmlEntity()->getOwner().compare(mNodeId) == 0)
+    {
+        LogManager::getSingleton().logMessage("AvatarNode::addAwareEntity() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " owned entity uid:" + StringConverter::toString(entity->getXmlEntity()->getUid()));
         mOwnedEntities[entity->getXmlEntity()->getUid()] = entity;
+    }
     else
+    {
+        LogManager::getSingleton().logMessage("AvatarNode::addAwareEntity() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " adding listener on entity uid:" + StringConverter::toString(entity->getXmlEntity()->getUid()));
         entity->addEntityListener(this);
+    }
 
 #ifdef PHYSICSPLUGINS
     // for instance we create physics of scenes + my own avatar
     if (!((entity->getXmlEntity()->getType() == ETAvatar) && (entity->getXmlEntity()->getOwner().compare(mNodeId) != 0)))
     {
+        LogManager::getSingleton().logMessage("AvatarNode::addAwareEntity() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " creating physics of entity uid:" + StringConverter::toString(entity->getXmlEntity()->getUid()));
         entity->createPhysics(mPhysicsScene);
         if (entity->getXmlEntity()->getType() == ETSite)
             mAvatar.setGravity(true);
@@ -85,11 +104,20 @@ bool AvatarNode::addAwareEntity(Entity* entity)
 
     pthread_mutex_unlock(&mMutex);
 
+#ifdef POOL
+    RefCntPoolPtr<XmlEvt> evt;
+    evt->setType(ETNewEntity);
+    evt->setDatas(RefCntPoolPtr<XmlData>(entity->getXmlEntity()));
+    pthread_mutex_lock(&mEvtsMutex);
+    mEvtsToHandleList.push_back(evt);
+    pthread_mutex_unlock(&mEvtsMutex);
+#else
     XmlEvt* evt = new XmlEvt(ETNewEntity);
     evt->setDatas(entity->getXmlEntity());
     pthread_mutex_lock(&mEvtsMutex);
     mEvtsToHandleList.push_back(evt);
     pthread_mutex_unlock(&mEvtsMutex);
+#endif
 
     return true;
 }
@@ -97,24 +125,42 @@ bool AvatarNode::addAwareEntity(Entity* entity)
 //-------------------------------------------------------------------------------------
 bool AvatarNode::removeAwareEntity(Entity* entity)
 {
+    LogManager::getSingleton().logMessage("AvatarNode::removeAwareEntity() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " removing entity uid:" + StringConverter::toString(entity->getXmlEntity()->getUid()));
+
     pthread_mutex_lock(&mMutex);
 
     if (entity->getXmlEntity()->getOwner().compare(mNodeId) != 0)
+    {
+        LogManager::getSingleton().logMessage("AvatarNode::removeAwareEntity() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " removing listener on entity uid:" + StringConverter::toString(entity->getXmlEntity()->getUid()));
         entity->removeEntityListener(this);
+    }
 
     mAwareEntities.erase(entity->getXmlEntity()->getUid());
 
 #ifdef PHYSICSPLUGINS
-    entity->destroyPhysics();
+    if (!((entity->getXmlEntity()->getType() == ETAvatar) && (entity->getXmlEntity()->getOwner().compare(mNodeId) != 0)))
+    {
+        LogManager::getSingleton().logMessage("AvatarNode::removeAwareEntity() uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " destroying physics of entity uid:" + StringConverter::toString(entity->getXmlEntity()->getUid()));
+        entity->destroyPhysics();
+    }
 #endif
 
     pthread_mutex_unlock(&mMutex);
 
+#ifdef POOL
+    RefCntPoolPtr<XmlEvt> evt;
+    evt->setType(ETLostEntity);
+    evt->setDatas(RefCntPoolPtr<XmlData>(entity->getXmlEntity()));
+    pthread_mutex_lock(&mEvtsMutex);
+    mEvtsToHandleList.push_back(evt);
+    pthread_mutex_unlock(&mEvtsMutex);
+#else
     XmlEvt* evt = new XmlEvt(ETLostEntity);
     evt->setDatas(entity->getXmlEntity());
     pthread_mutex_lock(&mEvtsMutex);
     mEvtsToHandleList.push_back(evt);
     pthread_mutex_unlock(&mEvtsMutex);
+#endif
 
     return true;
 }
@@ -126,7 +172,11 @@ bool AvatarNode::processEvt(XmlEvt& xmlEvt, std::string& xmlRespStr)
     static unsigned long l = (unsigned long)-1;
     if (xmlEvt.getType() == ETUpdatedEntity)
     {
+#ifdef POOL
+        XmlEntity* xmlEntity = (XmlEntity*)xmlEvt.getDatas().get();
+#else
         XmlEntity* xmlEntity = (XmlEntity*)xmlEvt.getDatas();
+#endif
         if (xmlEntity == 0)
         {
             xmlRespStr = "No entity found in event !";
@@ -175,11 +225,20 @@ bool AvatarNode::processEvt(XmlEvt& xmlEvt, std::string& xmlRespStr)
 }
 
 //-------------------------------------------------------------------------------------
+#ifdef POOL
+bool AvatarNode::freeEvt(RefCntPoolPtr<XmlEvt>& evt)
+#else
 bool AvatarNode::freeEvt(XmlEvt* evt)
+#endif
 {
     pthread_mutex_lock(&mEvtsMutex);
+#ifdef POOL
+    if (evt->getDatas() == mAvatar.mUpdatedXmlEntity)
+        mAvatar.mUpdatedXmlEntity->setDefinedAttributes(XmlEntity::DANone);
+#else
     if (evt->getDatas() == &mAvatar.mUpdatedXmlEntity)
         mAvatar.mUpdatedXmlEntity.setDefinedAttributes(XmlEntity::DANone);
+#endif
     pthread_mutex_unlock(&mEvtsMutex);
 
     return Node::freeEvt(evt);
@@ -231,6 +290,18 @@ c++;
     if (mAvatar.mDirty)
     {
         pthread_mutex_lock(&mEvtsMutex);
+#ifdef POOL
+        if (!mAvatar.mUpdatedXmlEntity->getDefinedAttributes() & XmlEntity::DAUid)
+        {
+            RefCntPoolPtr<XmlEvt> evt;
+            evt->setType(ETUpdatedEntity);
+            mAvatar.mUpdatedXmlEntity->setUid(mAvatar.getXmlEntity()->getUid());
+            mAvatar.mUpdatedXmlEntity->setPosition(mAvatar.getXmlEntity()->getPosition());
+#ifdef LOGSNDRCV
+            LogManager::getSingleton().logMessage("SND uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " " + StringConverter::toString(mAvatar.getXmlEntity()->getPosition()));
+#endif
+            evt->setDatas(RefCntPoolPtr<XmlData>(mAvatar.mUpdatedXmlEntity));
+#else
         if (!mAvatar.mUpdatedXmlEntity.getDefinedAttributes() & XmlEntity::DAUid)
         {
             XmlEvt* evt = new XmlEvt(ETUpdatedEntity);
@@ -240,6 +311,7 @@ c++;
             LogManager::getSingleton().logMessage("SND uid:" + StringConverter::toString(mAvatar.getXmlEntity()->getUid()) + " " + StringConverter::toString(mAvatar.getXmlEntity()->getPosition()));
 #endif
             evt->setDatas(&mAvatar.mUpdatedXmlEntity);
+#endif
             mEvtsToHandleList.push_back(evt);
             mAvatar.throwUpdateToEntityListeners(*this, mAvatar, *evt);
             mAvatar.mDirty = false;
@@ -259,6 +331,21 @@ bool AvatarNode::updated(const Node& node, Entity& entity, XmlEvt& xmlEvt)
 //    XmlEvt* evt = new XmlEvt(ETUpdatedEntity);
 //    evt->setDatas(xmlEvt.getDatas());
 //    mEvtsToHandleList.push_back(evt);
+#ifdef POOL
+    if (mXmlEntityMap.find(entity.getXmlEntity()->getUid()) == mXmlEntityMap.end())
+        mXmlEntityMap[entity.getXmlEntity()->getUid()] = RefCntPoolPtr<XmlEntity>(new XmlEntity(entity.getXmlEntity()->getUid()));
+    RefCntPoolPtr<XmlEntity>& local = mXmlEntityMap[entity.getXmlEntity()->getUid()];
+    XmlEntity* evtEntity = (XmlEntity*)xmlEvt.getDatas().get();
+    local->setDefinedAttributes(evtEntity->getDefinedAttributes());
+    if (local->getDefinedAttributes() & XmlEntity::DAPosition)
+        local->setPosition(evtEntity->getPosition());
+    if (local->getDefinedAttributes() & XmlEntity::DAOrientation)
+        local->setOrientation(evtEntity->getOrientation());
+    RefCntPoolPtr<XmlEvt> evt;
+    evt->setType(ETUpdatedEntity);
+    evt->setDatas(RefCntPoolPtr<XmlData>(local));
+    mEvtsToHandleList.push_back(evt);
+#else
     if (mXmlEntityMap.find(entity.getXmlEntity()->getUid()) == mXmlEntityMap.end())
         mXmlEntityMap[entity.getXmlEntity()->getUid()] = new XmlEntity(entity.getXmlEntity()->getUid());
     XmlEntity* local = mXmlEntityMap[entity.getXmlEntity()->getUid()];
@@ -270,6 +357,7 @@ bool AvatarNode::updated(const Node& node, Entity& entity, XmlEvt& xmlEvt)
     XmlEvt* evt = new XmlEvt(ETUpdatedEntity);
     evt->setDatas(local);
     mEvtsToHandleList.push_back(evt);
+#endif
     pthread_mutex_unlock(&mEvtsMutex);
 
     return true;
